# node_gateway/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CASE10 Node API Gateway",
    "version": "0.1.0",
    "description": "Minimal OpenAPI contract for the CASE10 Node.js gateway. The gateway proxies CASE10 requests to the existing Python API engine."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Local docker-compose gateway"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "Login helper for gateway smoke and clients",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token"
          },
          "401": {
            "$ref": "#/components/responses/PythonError"
          }
        }
      }
    },
...
    "/api/documents": {
      "get": {
        "summary": "List uploaded documents for a project",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectIdQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Document list"
          }
        }
      }
    },
    "/api/documents/{document_id}": {
      "delete": {
        "summary": "Delete an uploaded document",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          },
          {
            "$ref": "#/components/parameters/ProjectIdQuery"
          },
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "rag" }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result"
          },
          "404": {
            "$ref": "#/components/responses/PythonError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
...
      "ProjectIdQuery": {
        "name": "project_id",
        "in": "query",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
...
    },
    "responses": {
      "PythonError": {
        "description": "HTTP error body forwarded from Python API",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": [
          "login",
          "password"
        ],
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {}
        }
      }
    }
  }
}